requirements.txt
# Django系統ライブラリ
Django==5.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-import-export==3.3.5
djangorestframework==3.14.0

# DB接続ライブラリ
mysqlclient==2.2.1

# コード整形ライブラリ
ruff==0.4.7